Move box shadow building methods to box_shadow.rs.
